int/web/register: renames package.
