+++ backend/internal/web/register/agent.go
@@ -1,4 +1,4 @@
-package registration
+package register
 
 import (
 	"fmt"
